Now it looks very much like venv
workspace.go will create the folder where the new gopath will be.
We need the gopath to be different from the golang binary path
per Golang requirements. The golang path will already have a src
and bin directory. "go get" should be able to operate in this
environment since there is a "src" directory". It will also
write the "activate" script in the required path i.e. the one
specified in the command line arguments.

script.go will create the script. I made some changes so it will
display in the prompt that it is working with the new environment.
variable for GOPATH.

Currently, this works *only* for GOPATH, but that's a pretty good
start!

Signed-off-by: nisha <nishakumarx@gmail.com>