Add switching go versions
I can now do:
genv -version <other go version> <dir>
cd <dir>
source bin/activate

and I should be able to get `go` to point to the correct go
binary.

All this does is use "which" to get the binary path and then
make a symlink pointing to that path within the bin directory
of the given directory.

Signed-off-by: nisha <nishakumarx@gmail.com>